refactor: Move auth related code to utils subfolder.
