+++ app/routes/_foh+/_auth+/login.tsx
@@ -1,8 +1,12 @@
 import type { ActionFunctionArgs, LoaderFunctionArgs } from '@remix-run/node';
 import { json, useLoaderData, useSubmit } from '@remix-run/react';
 
-import { authenticator, commitSession, getSession } from '#.server/auth';
 import { Button, Form, TextField } from '#components/(ui)';
+import {
+  authenticator,
+  commitSession,
+  getSession,
+} from '#utils/auth/auth.server';
 
 export async function loader({ request }: LoaderFunctionArgs) {
   await authenticator.isAuthenticated(request, {
