+++ app/routes/_foh+/_auth+/logout.tsx
@@ -1,5 +1,5 @@
 import type { LoaderFunctionArgs } from '@remix-run/node';
-import { authenticator } from '#.server/auth';
+import { authenticator } from '#utils/auth/auth.server';
 
 export async function loader({ request }: LoaderFunctionArgs) {
   return await authenticator.logout(request, {
