+++ app/routes/_foh+/_auth+/onboarding.tsx
@@ -2,7 +2,11 @@ import type { ActionFunctionArgs, LoaderFunctionArgs } from '@remix-run/node';
 import { json, redirect, useLoaderData, useSubmit } from '@remix-run/react';
 import { Button, Form, TextField } from '#components/(ui)';
 
-import { authenticator, commitSession, getSession } from '#.server/auth';
+import {
+  authenticator,
+  commitSession,
+  getSession,
+} from '#utils/auth/auth.server';
 
 export async function loader({ request }: LoaderFunctionArgs) {
   await authenticator.isAuthenticated(request, {
