+++ app/routes/manager+/_layout/_layout.tsx
@@ -1,7 +1,7 @@
 import type { LoaderFunctionArgs } from '@remix-run/node';
 import { Outlet } from '@remix-run/react';
-import { requireAdmin } from '#.server/auth';
 import { ThemeMenu } from '#components';
+import { requireAdmin } from '#utils/auth/auth.server';
 import { Sidebar } from './sidebar';
 
 export async function loader({ request }: LoaderFunctionArgs) {
